# MM-42810: Introduce a channel hook for a websocket event