… server/channels/app/authorization_test.go
… server/channels/app/platform/web_hub_test.go
… server/channels/app/post_metadata_test.go