@lieut-data 2023-06-21
> For this issue, the main problem was that since we don't know
which users have permissions to which channels, we had to go through
all members of a channel to figure that out. This was redundant
since a large portion of those users might not even be connected
at that time.

Another question: if most of the channel members are connected, do we still glean any savings? The issue in question primarily arose for me from permalink posts in a channel like town square, and I'm wondering if even having 30-40% of the members connected means we'll still have the fundamental issue at play here.