@agnivade 2023-06-21
> Another question: if most of the channel members are connected, do we still glean any savings? 

We don't. And that's where the last line of my PR comes into play ;)

> The next iteration of this should be to use websocket subscriptions
to monitor exactly which users are on that channel to even
trim down that list. That is a larger initiative to be taken later.