@lieut-data 2023-06-21
> The next iteration of this should be to use websocket subscriptions
to monitor exactly which users are on that channel to even
trim down that list.

Would the idea be to not send these payloads if the users are simply looking at a different channel? Would that case a thundering herd of users switch to an `@here` in `~town-square` and trying to resolve the permalink preview via an API request?