@agnivade 2023-06-22
@lieut-data - PTAL

 - Removed unnecessary messageCopy
  - Reworked sanitizePostMetadataForUserAndChannel to be used from the hook
  so that unit test coverage exists.
  - Also modified it to return a true/false rather than actually modify the post.
  This simplifies the data race issues while handling the post metadata. Since
  we just have to deal with metadata on/off, we cache the modified JSON with no metadata
  and just check whether the user has access or not. If they don't, then we attach the
  modified JSON. This also avoids the need to JSON marshal every time.
  - Added more comments to explain what's going on.