@hmhealey 2023-06-22
I'm even less than 0/5 on this since I'm proposing an entirely different way of supporting this case, but what if, instead of sanitizing the websocket event for certain users, we instead extend the `broadcast` parameter to be allow for more control over the event and then we send a sanitized and unsanitized one slightly different `broadcast` settings? Like for this case, we could prepare two messages: 

1. The first one would have the post attached and be sent to users who are both in the post's channels and the permalinked channel.
2. The second one wouldn't have the post attached and would be send to users who are in the post's channel but not the permalinked channel.

That obviously requires extra work to support logical rules, and we'd have to make sure to maintain compatibility with clients that rely on the `broadcast` option in its current state, but that feels like it would be easier to reason about compared to having to understand when the event gets sanitized.

Feel free to ignore this though since I can't stop myself from trying to find different ways to do things that already have good solution 😛