@agnivade 2023-06-23
@streamer45 @lieut-data - LMK how does this new iteration look:

- Made the check inline
- Invert the check in the hook to be more security friendly
- Changing publish event to do everything
- Added test for HasPermissionToReadChannel
- Added test for the channel hook