@agnivade 2023-06-23
>  Like for this case, we could prepare two messages:
The first one would have the post attached and be sent to users who are both in the post's channels and the permalinked channel.
The second one wouldn't have the post attached and would be send to users who are in the post's channel but not the permalinked channel.

@hmhealey - Can you expand further on the implementation of this? How will you find out which connected users have permission to the permalinked channel _before_ entering the websocket layer? And if you don't then you get into the same problem.

Therefore to know which users are connected, you need to enter the websocket layer, and then you will need some sort of hook at that point to check if the user has the permission or not.

Let me know.