@hmhealey 2023-06-23
I'm sort of meaning that we could eventually change the Broadcast field to more or less give it some basic boolean logic for things like filtering channel IDs. Sketching something out, it might look like
```go
func SendCreatePostEvents(post *model.Post, permalinkedPost *model.Post) {
	if permalinkedPost != nil {
		send(&model.WebSocketMessage{
			Type: model.WebsocketEventPosted,
			Broadcast: {
				Filter: And(
					MemberOfChannel(post.ChannelId),
					NotMemberOfChannel(permalinkedPost.ChannelId),
				),
			},
			Data: {
				Post: post,
			},
		})

		send(&model.WebSocketMessage{
			Type: model.WebsocketEventPosted,
			Broadcast: {
				Filter: And(
					MemberOfChannel(post.ChannelId),
					MemberOfChannel(permalinkedPost.ChannelId),
				),
			},
			Data: {
				Post: post.WithEmbeddedPostMetadata(permalinkedPost),
			},
		})
	} else {
		send(&model.WebSocketMessage{
			Type: model.WebsocketEventPosted,
			Broadcast: {
				Filter: MemberOfChannel(post.ChannelId),
			},
			Data: {
				Post: post,
			},
		})
	}
}
```

We'd still have the WebSocket code itself handling permissions checking, and it could do it in a smarter way like you're doing in this PR, but I think this could be a nice future change to make it very clear to someone reading the code who should receive which message. As it is, we're kind of limited by the structure of `Broadcast` which is what makes these cases hard like with sanitizing for non-admins or having more specific messages for posts with permalink previews.

But like I said, that's entirely a different direction than any of our existing code, and it would require rewriting a ton of stuff, so I was pitching it as a totally hypothetical thing