@mattermost-build 2023-06-27
You don't have permissions to trigger this command.
 It's only available for organization members.