@agnivade 2023-06-27
Hi @DHaussermann - this needs some testing around permalink previews. Steps are mentioned in the JIRA ticket. Let me know if you have any questions.  Thanks!