@mattermost-build 2023-06-27
Successfully triggered E2E testing!
[GitLab pipeline](https://git.internal.mattermost.com/qa/cypress-ui-automation/-/pipelines/348440) | [Test dashboard](https://automation-dashboard.vercel.app/cycle/348440)